Add files via upload
This zipped file contains the complete information about Echo- music player which is an android application. It fetches the songs from the Android media offline. Also has an exciting feature "shake to change the song" which changes to next song when this setting is applied through settings in the application. It also has a favourite page where all the favourite songs are mentioned.